Initial Pass at the Script
Creates a script that uses a token to authenticate uploads and proxies
the images back to the browser.

The token is posted as part of the form. I did this because reading
headers relies on a specific web server API to access (e.g. Apache). So
it seemed like it would be more consistent to just use a form field.

The UPLOAD_DIR does not have it's content served publicly like I
originally planned. Instead the script will use `readfile` to proxy it
back to the requester. This allows that directory to be located out of
the website and prevents scripts from being run.

Configuration can be handled by environment variables, but doesn't have to
be. There is a way to modify this script to set the token and upload
directory path.

I used the `global` keyword a few times. Pretty sure this is a Bad
Idea™, but it seemed like a good idea at the time. What could go wrong?